requirements.txt
streamlit>=1.28.0
requests>=2.31.0
openai>=1.3.0
sentence-transformers>=2.2.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.17.0
hdbscan>=0.8.33
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
nltk>=3.8.0
transformers>=4.36.0
torch>=2.1.0
